Update token_json.dart to have num value for supply parameter. Extract gradient avatar to a widget. Fix height in wallet_details_page.dart. Remove sized box from wallets_list_navigator.dart

+++ lib/data/model/translators/prices_translator.dart
@@ -16,7 +16,7 @@ extension PricesTranslator on PricesDataJson {
                   (e) => Token(
                     denom: Denom(e.symbol),
                     amount: Amount.fromString(e.price.toString()),
-                    supply: e.supply,
+                    supply: e.supply.toDouble(),
                   ),
                 )
                 .toList() ??
